Docker setup is not maintained, but may still work
This commit is designed only to add a comment on the github page so that
anyone who clones/forks this repo knows that the docker set up is not in
active development.  It was included by the original author (goropikari)
but was never tested before or after changes by the second author
(tlesaul2).